requirements.txt
PyYAML~=6.0
Flask~=2.2
requests~=2.28
gunicorn~=23.0
